# api/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Open Insurance API
  description: Modernized Insurance API
  contact:
    email: connect@openinsurance.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.0.9999
servers:
- url: https://virtserver.swaggerhub.com/JimmysOrg/openinsurance/0.0.0.9999
  description: SwaggerHub API Auto Mocking
...
paths:
  /vehicle:
    get:
      tags:
      - developers
      summary: returns vehicle
      description: |
        By passing in the appropriate options, you can search for
        Vehicle in the system
      operationId: search_vehicle_by_vehicle_id
      parameters:
      - name: searchString
        in: query
        description: pass an optional search string for looking up Vehicle
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: skip
        in: query
        description: number of records to skip for pagination
        required: false
        style: form
        explode: true
        schema:
          minimum: 0
          type: integer
          format: int32
      - name: limit
        in: query
        description: maximum number of records to return
        required: false
        style: form
        explode: true
        schema:
          maximum: 50
          minimum: 0
          type: integer
          format: int32
      responses:
        "200":
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vehicle'
                x-content-type: application/json
        "400":
          description: bad input parameter
      x-openapi-router-controller: swagger_server.controllers.developers_controller
    post:
      tags:
      - admins
      summary: adds a vehicle
      description: Adds an item to the system
      operationId: add_vehicle
      requestBody:
        description: Vehicle to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vehicle'
      responses:
        "201":
          description: item created
        "400":
          description: "invalid input, object invalid"
        "409":
          description: an existing item already exists
      x-openapi-router-controller: swagger_server.controllers.admins_controller
  /claim:
    get:
      tags:
      - developers
      summary: returns claim
      description: |
        By passing in the appropriate options, you can search for
        claim in the system
      operationId: search_claim_by_claim_id
      parameters:
      - name: searchString
        in: query
        description: pass an optional search string for looking up claim
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: skip
        in: query
        description: number of records to skip for pagination
        required: false
        style: form
        explode: true
        schema:
          minimum: 0
          type: integer
          format: int32
      - name: limit
        in: query
        description: maximum number of records to return
        required: false
        style: form
        explode: true
        schema:
          maximum: 50
          minimum: 0
          type: integer
          format: int32
      responses:
        "200":
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Claim'
                x-content-type: application/json
        "400":
          description: bad input parameter
      x-openapi-router-controller: swagger_server.controllers.developers_controller
    post:
      tags:
      - admins
      summary: adds a claim
      description: Adds an item to the system
      operationId: add_claim
      requestBody:
        description: Claim to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Claim'
      responses:
        "201":
          description: item created
        "400":
          description: "invalid input, object invalid"
        "409":
          description: an existing item already exists
      x-openapi-router-controller: swagger_server.controllers.admins_controller
  /driver:
    get:
      tags:
      - developers
      summary: returns driver
      description: |
        By passing in the appropriate options, you can search for
        claim in the system
      operationId: search_driver_by_driver_id
      parameters:
      - name: searchString
        in: query
        description: pass an optional search string for looking up driver
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: skip
        in: query
        description: number of records to skip for pagination
        required: false
        style: form
        explode: true
        schema:
          minimum: 0
          type: integer
          format: int32
      - name: limit
        in: query
        description: maximum number of records to return
        required: false
        style: form
        explode: true
        schema:
          maximum: 50
          minimum: 0
          type: integer
          format: int32
      responses:
        "200":
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Driver'
                x-content-type: application/json
        "400":
          description: bad input parameter
      x-openapi-router-controller: swagger_server.controllers.developers_controller
    post:
      tags:
      - admins
      summary: adds a Driver
      description: Adds an item to the system
      operationId: add_driver
      requestBody:
        description: Driver to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Driver'
      responses:
        "201":
          description: item created
        "400":
          description: "invalid input, object invalid"
        "409":
          description: an existing item already exists
      x-openapi-router-controller: swagger_server.controllers.admins_controller
  /motorCoverage:
    get:
      tags:
      - developers
      summary: returns motorCoverage
      description: |
        By passing in the appropriate options, you can search for
        claim in the system
      operationId: search_motor_coverage_by_motor_coverage_id
      parameters:
      - name: searchString
        in: query
        description: pass an optional search string for looking up motorCoverage
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: skip
        in: query
        description: number of records to skip for pagination
        required: false
        style: form
        explode: true
        schema:
          minimum: 0
          type: integer
          format: int32
      - name: limit
        in: query
        description: maximum number of records to return
        required: false
        style: form
        explode: true
        schema:
          maximum: 50
          minimum: 0
          type: integer
          format: int32
      responses:
        "200":
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Driver'
                x-content-type: application/json
        "400":
          description: bad input parameter
      x-openapi-router-controller: swagger_server.controllers.developers_controller
    post:
      tags:
      - admins
      summary: adds a MotorCoverage
      description: Adds an item to the system
      operationId: add_motor_coverage
      requestBody:
        description: motorCoverage to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/motorCoverage'
      responses:
        "201":
          description: item created
        "400":
          description: "invalid input, object invalid"
        "409":
          description: an existing item already exists
      x-openapi-router-controller: swagger_server.controllers.admins_controller
components:
  schemas:
    Vehicle:
      type: object
      properties:
        id:
          type: integer
          format: int64
        HP:
          type: number
          format: float
        KW:
          type: number
          format: float
        torque:
          type: number
          format: float
        evHP:
          type: number
          format: float
        evKW:
          type: number
          format: float
        evTorque:
          type: number
          format: float
        acceleration:
          type: number
          format: float
        vehicleTopSpeed:
          type: number
          format: float
        engineCapacity:
          type: number
          format: float
        modelYear:
          type: number
          format: float
        hasTractionEnabled:
          type: boolean
        hasImmobilizer:
          type: boolean
        hasTheftDetection:
          type: boolean
        currentMileageDynamic:
          type: number
          format: float
        yearlyMilageDynamic:
          type: number
          format: float
        highwayYearlyMilageDynamic:
          type: number
          format: float
        dailyMilageDynamic:
          type: number
          format: float
        serviceHistory:
          type: number
          format: float
        serviceDue:
          type: number
          format: float
        timeToService:
          type: number
          format: float
        recallHistory:
          type: number
          format: float
        tractionControlEngaged:
          type: boolean
        accelerationLongitudinal:
          type: number
          format: float
        accelerationLateral:
          type: number
          format: float
        accelerationVertical:
          type: number
          format: float
        brakingFrequency:
          type: number
          format: float
        brakePedalForce:
          type: number
          format: float
        brakePedalSpeed:
          type: number
          format: float
        performanceMode:
          type: number
          format: float
        emergencyBraking:
          type: number
          format: float
        engnitionOn:
          type: boolean
        engnitionOff:
          type: boolean
        ignitionOnTime:
          type: number
          format: float
        ignitionOffTime:
          type: number
          format: float
        logitude:
          type: number
          format: float
        latitude:
          type: number
          format: float
        altitude:
          type: number
          format: float
        heading:
          type: number
          format: float
        isMoving:
          type: boolean
        hornIsActive:
          type: boolean
        drivingSpeed:
          type: number
          format: float
        wheelSpin:
          type: number
          format: float
        decelrationRate:
          type: number
          format: float
        steeringSpeedTurn:
          type: number
          format: float
        laneDepartureWarnning:
          type: number
          format: float
        adasAbsIsActive:
          type: boolean
        obstacleDetectionIsActive:
          type: boolean
        driverIntervention:
          type: boolean
        obstacleDetectionWarning:
          type: string
        speedSet:
          type: number
          format: float
        yaw:
          type: number
          format: float
        pitch:
          type: number
          format: float
        roll:
          type: number
          format: float
        gForce:
          type: number
          format: float
        trailerIncluded:
          type: boolean
        row1Pos1Isbelted:
          type: boolean
        row1Pos2Isbelted:
          type: boolean
        row2Pos1Isbelted:
          type: boolean
        row2Pos2Isbelted:
          type: boolean
        cabinTemp:
          type: number
          format: float
        cabinHumidity:
          type: number
          format: float
        tireConditionRow1Left:
          type: string
        tireConditionRow1Right:
          type: string
        tireConditionRow2Left:
          type: string
        tireConditionRow2Right:
          type: string
        tirePressureRow1Left:
          type: string
        tirePressureRow1Right:
          type: string
        tirePressureRow2Left:
          type: string
        tirePressureRow2Right:
          type: string
        brakePadWearRow1Left:
          type: string
        brakePadWearRow1Right:
          type: string
        brakePadWearRow2Left:
          type: string
        brakePadWearRow2Right:
          type: string
        brakesWearRow1Right:
          type: string
        brakesWearRow1Left:
          type: string
        brakeWearRow2Left:
          type: string
        brakeWearRow2Right:
          type: string
        clutchWear:
          type: number
          format: float
        dashboardWarnning:
          type: string
        hasAirbagDeployed:
          type: boolean
        hasBrakesError:
          type: boolean
        engineWarning:
          type: string
        KnownVehicleDamage:
          type: string
        damagedParts:
          type: string
        damagedPartsCost:
          type: number
          format: float
        upcomingService:
          type: string
          format: date
        occupiedSeats:
          type: integer
          format: int32
        childSeatOccupiesSeat:
          type: integer
          format: int32
        consentGranted:
          type: boolean
        plateNumber:
          type: string
        registrationDate:
          type: string
          format: date
        countryOfRegistration:
          type: string
        chassisNumber:
          type: string
        vin:
          type: string
        engineNumber:
          type: string
        vehicleWeight:
          type: number
          format: float
        agencyRepair:
          type: boolean
        bodyType:
          type: string
          format: bodyType
        fuelType:
          type: string
          format: fuelType
        aiClassification:
          type: string
          format: aiClassification
        vehicleUse:
          type: boolean
          format: vehicleUse
        vehicleGarage:
          type: boolean
        vehicleBrand:
          type: string
        vehicleModel:
          type: string
        seats:
          type: integer
          format: int32
        colour:
          type: string
        sumInsured:
          type: number
          format: float
        accessories:
          type: string
        accessoryValue:
          type: string
        co2Emissions:
          type: boolean
        automaticTransmission:
          type: boolean
        lefthandDrive:
          type: boolean
        doors:
          type: integer
          format: int32
        securityDevice:
          type: boolean
        modification:
          type: boolean
        digitalKeyUsed:
          type: boolean
        blueBadgeAdapted:
          type: boolean
        distanceUnit:
          type: string
        pleasureDistance:
          type: number
          format: float
        businessDistance:
          type: number
          format: float
        currentMileage:
          type: integer
          format: int32
        yearlyMilage:
          type: integer
          format: int32
      example:
        brakePadWearRow2Right: brakePadWearRow2Right
        distanceUnit: distanceUnit
        businessDistance: 4.678948
        engnitionOn: true
        decelrationRate: 3.0937452
        upcomingService: 2000-01-23
        blueBadgeAdapted: true
        HP: 6.0274563
        highwayYearlyMilageDynamic: 1.0246457
        tireConditionRow1Left: tireConditionRow1Left
        hasBrakesError: true
        cabinTemp: 7.7403517
        sumInsured: 6.628464
        gForce: 8.969579
        clutchWear: 3.05761
        brakePadWearRow1Right: brakePadWearRow1Right
        accelerationLongitudinal: 5.025005
        vin: vin
        id: 0
        pitch: 0.10263654
        occupiedSeats: 5
        KnownVehicleDamage: KnownVehicleDamage
        countryOfRegistration: countryOfRegistration
        ignitionOnTime: 6.965118
        modelYear: 4.145608
        currentMileageDynamic: 7.386282
        laneDepartureWarnning: 0.8851375
        hasTheftDetection: true
        cabinHumidity: 3.0205798
        doors: 4
        emergencyBraking: 3.5571952
        brakesWearRow1Left: brakesWearRow1Left
        damagedPartsCost: 7.0483656
        timeToService: 1.1730742
        latitude: 6.778325
        roll: 4.652396
        torque: 5.962134
        recallHistory: 4.9652185
        hasAirbagDeployed: true
        securityDevice: true
        tireConditionRow2Right: tireConditionRow2Right
        evHP: 5.637377
        brakePadWearRow1Left: brakePadWearRow1Left
        pleasureDistance: 1.0414449
        brakingFrequency: 6.6835623
        vehicleModel: vehicleModel
        tractionControlEngaged: true
        engineWarning: engineWarning
        lefthandDrive: true
        accelerationVertical: 9.36931
        ignitionOffTime: 1.284659
        brakePedalSpeed: 9.018348
        speedSet: 7.05877
        heading: 5.9448957
        tirePressureRow1Right: tirePressureRow1Right
        tirePressureRow2Left: tirePressureRow2Left
        engineNumber: engineNumber
        accessories: accessories
        vehicleUse: true
        KW: 1.4658129
        plateNumber: plateNumber
        engineCapacity: 2.027123
        yaw: 6.519181
        logitude: 2.8841622
        fuelType: fuelType
        yearlyMilage: 9
        performanceMode: 6.4384236
        obstacleDetectionWarning: obstacleDetectionWarning
        drivingSpeed: 6.704019
        aiClassification: aiClassification
        brakePedalForce: 8.762042
        brakeWearRow2Left: brakeWearRow2Left
        bodyType: bodyType
        currentMileage: 7
        evTorque: 7.0614014
        accelerationLateral: 9.965781
        seats: 0
        acceleration: 9.301444
        dailyMilageDynamic: 1.4894159
        accessoryValue: accessoryValue
        tirePressureRow1Left: tirePressureRow1Left
        isMoving: true
        registrationDate: 2000-01-23
        childSeatOccupiesSeat: 3
        tireConditionRow2Left: tireConditionRow2Left
        co2Emissions: true
        serviceHistory: 6.846853
        row1Pos2Isbelted: true
        brakePadWearRow2Left: brakePadWearRow2Left
        brakesWearRow1Right: brakesWearRow1Right
        modification: true
        wheelSpin: 3.3531933
        chassisNumber: chassisNumber
        brakeWearRow2Right: brakeWearRow2Right
        driverIntervention: true
        yearlyMilageDynamic: 1.2315135
        row2Pos2Isbelted: true
        consentGranted: true
        agencyRepair: true
        vehicleWeight: 4.078846
        altitude: 6.878052
        tirePressureRow2Right: tirePressureRow2Right
        hasTractionEnabled: true
        obstacleDetectionIsActive: true
        serviceDue: 7.4577446
        adasAbsIsActive: true
        vehicleBrand: vehicleBrand
        trailerIncluded: true
        hornIsActive: true
        tireConditionRow1Right: tireConditionRow1Right
        row1Pos1Isbelted: true
        hasImmobilizer: true
        row2Pos1Isbelted: true
        automaticTransmission: true
        damagedParts: damagedParts
        vehicleGarage: true
        evKW: 2.302136
        colour: colour
        dashboardWarnning: dashboardWarnning
        steeringSpeedTurn: 7.143538
        engnitionOff: true
        vehicleTopSpeed: 3.6160767
        digitalKeyUsed: true
      xml:
        name: Vehicle
    Claim:
      type: object
      properties:
        claimType:
          type: string
          format: claimType
        location:
          type: string
        lossCause:
          type: string
          format: perils
        description:
          type: string
        fnol:
          type: string
          format: date
        claimNumber:
          type: string
        liabilityShare:
          type: number
          format: float
        reserve:
          type: number
          format: float
        claimStatus:
          type: string
          format: claimStatus
        lastUpdate:
          type: string
          format: date
        reopenDate:
          type: string
          format: date
        excessAmount:
          type: number
          format: float
        paymentMethod:
          type: string
          format: paymentMethod
        documents:
          type: string
        lossDate:
          type: string
          format: date
      example:
        documents: documents
        liabilityShare: 0.8008282
        description: description
        excessAmount: 1.4658129
        lossCause: lossCause
        claimType: claimType
        reopenDate: 2000-01-23
        fnol: 2000-01-23
        lossDate: 2000-01-23
        lastUpdate: 2000-01-23
        reserve: 6.0274563
        paymentMethod: paymentMethod
        location: location
        claimNumber: claimNumber
        claimStatus: claimStatus
      xml:
        name: Claim
    Driver:
      type: object
      properties:
        name:
          type: string
        gender:
          type: string
          format: gender
        driverDOB:
          type: string
          format: date
        isPrimaryDriver:
          type: boolean
        licence:
          type: string
          format: drivingLicense
        noClaimsDiscount:
          type: number
          format: float
        conviction:
          type: string
          format: conviction
        medicalConditon:
          type: string
          format: medicalCondition
        loading:
          type: number
          format: float
        isBlueBadge:
          type: boolean
        nonMotorConviction:
          type: string
        workStatus:
          type: string
          format: workStatus
        occupation:
          type: string
      example:
        driverDOB: 2000-01-23
        workStatus: workStatus
        licence: licence
        occupation: occupation
        gender: gender
        isPrimaryDriver: true
        conviction: conviction
        loading: 6.0274563
        isBlueBadge: true
        nonMotorConviction: nonMotorConviction
        name: name
        medicalConditon: medicalConditon
        noClaimsDiscount: 0.8008282
      xml:
        name: Driver
    motorCoverage:
      type: object
      properties:
        policyNumber:
          type: string
        inceptionDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        status:
          type: string
          format: policyStatus
        discountAmount:
          type: number
          format: float
        premiumRate:
          type: number
          format: float
        grossWrittenPremium:
          type: number
          format: float
        salesTax:
          type: number
          format: float
        brokeragePercentage:
          type: number
          format: float
        brokerageAmount:
          type: number
          format: float
        premiumPaymentFrequency:
          type: number
          format: premiumPaymentFrequency
        indemnityLimitPolicy:
          type: number
          format: int32
        indemnityLimitAccident:
          type: number
          format: int32
        isAgreedValue:
          type: boolean
        endorsementID:
          type: string
        endorsementDate:
          type: string
          format: date
        endorsementType:
          type: string
          format: endorsementType
        peril:
          type: string
          format: motorPeril
        voluntaryDeductiblePercentage:
          type: number
          format: float
        voluntaryDeductibleAmount:
          type: number
          format: int32
        compulsoryDeductiblePercentage:
          type: number
          format: float
        compulsoryDeductibleAmount:
          type: number
          format: int32
        windscreenDeductiblePercentage:
          type: number
          format: float
        windscreenDeductibleAmount:
          type: number
          format: int32
        distanceUnit:
          type: number
          format: distanceUnit
        pleasureDistance:
          type: number
          format: float
        businessDistance:
          type: number
          format: float
        numberOfVehicles:
          type: number
          format: int32
      xml:
        name: motorCoverage
...